Yoav's clarifications
From Yoav:

> Mempool ID. EIP-4337 supports the use of an alternate mempool that bundlers are able to opt into which is seperate from the canonical mempool.

To avoid collisions between groups of bundlers setting alt mempools, we can define a rule for the mempool ID.  I suggest keccak that covers the chain ID, EntryPoint address (to support multiple versions during transitions), and some format for describing the exceptions (e.g. a list of parameters such as max validation gas, whitelisted paymasters, etc.), or maybe a hash of a github link to a PR made against the specs repo to describe the rules of the alt mempool.  Something to make it unique.

> All bundler implementations should pass all tests in the Infinitism test suite, but don’t necessarily need to in order to participate in the bundler mempool. Need an integration test for p2p protocol.

I meant that there may be specific rules that are optional because they don't break the mempool apart (just put the bundler at risk), but most of the rules are implicitly mandatory because different implementations will result in bundlers rejecting each other as spammers.  E.g. if a bundler doesn't pass one of the opcode banning rules, then a UserOp that uses this opcode will cause this bundler to be kicked out by the other bundlers.